3596 - useUpdateDataSources: Move hook to parent
